General improvements. Improved storage managing. Merged wipe/detach methods. Release uri permission on detach. Fixed Storage import (checking for same ID doesn't make sense) Renamed editor paramter to "initialStorageId" to make this behavior more obvious.

+++ app/src/main/java/eu/darken/bb/storage/core/local/LocalStorage.kt
@@ -288,18 +288,13 @@ class LocalStorage @AssistedInject constructor(
                 return@map specInfo.copy(backups = newMetaData)
             }
 
-    // TODO Maybe release permission?
-    override fun detach(): Completable = Completable
-            .complete()
-            .doOnSubscribe { Timber.i("Detaching %s", storageRef) }
-
-    // TODO call detach after wipe?
-    override fun wipe(): Completable = Completable
+    override fun detach(wipe: Boolean): Completable = Completable
             .fromCallable {
-                storageRef.path.asFile().deleteAll()
+                if (wipe) storageRef.path.asFile().deleteAll()
             }
-            .doOnSubscribe { Timber.w("wipe().doOnSubscribe %s", storageRef) }
-            .doFinally { Timber.w("wipe().dofinally%s", storageRef) }
+            .doOnSubscribe { Timber.w("detach(wipe=%b).doOnSubscribe %s", wipe, storageRef) }
+            .doFinally { Timber.w("detach(wipe=%b).dofinally %s", wipe, storageRef) }
+
 
     private fun getSpecDir(specId: BackupSpec.Id): File = File(dataDir, specId.value)
